requirements.txt
openai>=1.40.0
feedparser>=6.0.11
requests>=2.31.0
python-dateutil>=2.9.0.post0
pytz>=2024.1
